+++ gamestonk_terminal/cryptocurrency/crypto_controller.py
@@ -9,15 +9,14 @@ from colorama import Style
 from prompt_toolkit.completion import NestedCompleter
 from binance.client import Client
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     EXPORT_BOTH_RAW_DATA_AND_FIGURES,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
-    menu_decorator,
     parse_known_args_and_warn,
     check_positive,
     system_clear,
-    try_except,
 )
 from gamestonk_terminal.menu import session
 from gamestonk_terminal.cryptocurrency.due_diligence import (
