+++ gamestonk_terminal/cryptocurrency/defi/defipulse_model.py
@@ -5,7 +5,7 @@ import requests
 from bs4 import BeautifulSoup
 import pandas as pd
 
-from gamestonk_terminal.helper_funcs import try_except
+from gamestonk_terminal.decorators import try_except
 
 
 @try_except
