+++ gamestonk_terminal/cryptocurrency/discovery/discovery_controller.py
@@ -5,13 +5,12 @@ __docformat__ = "numpy"
 import argparse
 from typing import List, Union
 from prompt_toolkit.completion import NestedCompleter
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     EXPORT_ONLY_RAW_DATA_ALLOWED,
-    menu_decorator,
     parse_known_args_and_warn,
     check_positive,
-    try_except,
     system_clear,
 )
 from gamestonk_terminal.menu import session
