+++ gamestonk_terminal/cryptocurrency/due_diligence/dd_controller.py
@@ -9,6 +9,7 @@ from colorama.ansi import Style
 import pandas as pd
 from binance.client import Client
 from prompt_toolkit.completion import NestedCompleter
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal.cryptocurrency.due_diligence import (
     coinglass_model,
     glassnode_model,
@@ -30,10 +31,8 @@ from gamestonk_terminal.cryptocurrency.due_diligence import (
 from gamestonk_terminal.helper_funcs import (
     EXPORT_BOTH_RAW_DATA_AND_FIGURES,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
-    menu_decorator,
     parse_known_args_and_warn,
     check_positive,
-    try_except,
     system_clear,
     valid_date,
 )
