+++ gamestonk_terminal/cryptocurrency/nft/nft_controller.py
@@ -2,13 +2,12 @@ import argparse
 from typing import List, Union
 
 from prompt_toolkit.completion import NestedCompleter
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.menu import session
 from gamestonk_terminal.helper_funcs import (
     EXPORT_ONLY_RAW_DATA_ALLOWED,
-    try_except,
     system_clear,
-    menu_decorator,
     check_positive,
     parse_known_args_and_warn,
 )
