+++ gamestonk_terminal/cryptocurrency/onchain/onchain_controller.py
@@ -11,6 +11,7 @@ from typing import List, Union
 from colorama.ansi import Style
 from prompt_toolkit.completion import NestedCompleter
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal.cryptocurrency.due_diligence.glassnode_model import (
     GLASSNODE_SUPPORTED_HASHRATE_ASSETS,
     INTERVALS,
@@ -23,11 +24,9 @@ from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.menu import session
 from gamestonk_terminal.helper_funcs import (
     EXPORT_BOTH_RAW_DATA_AND_FIGURES,
-    menu_decorator,
     parse_known_args_and_warn,
     check_positive,
     check_int_range,
-    try_except,
     system_clear,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
     valid_date,
