+++ gamestonk_terminal/cryptocurrency/overview/overview_controller.py
@@ -6,12 +6,11 @@ import argparse
 import difflib
 from typing import List, Union
 from prompt_toolkit.completion import NestedCompleter
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
-    menu_decorator,
     parse_known_args_and_warn,
     check_positive,
-    try_except,
     system_clear,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
 )
