+++ gamestonk_terminal/cryptocurrency/prediction_techniques/pred_controller.py
@@ -8,15 +8,14 @@ import pandas as pd
 import numpy as np
 from prompt_toolkit.completion import NestedCompleter
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
-    menu_decorator,
     parse_known_args_and_warn,
     check_positive,
     valid_date,
     get_next_stock_market_days,
     EXPORT_ONLY_FIGURES_ALLOWED,
-    try_except,
     system_clear,
 )
 from gamestonk_terminal.menu import session
