+++ gamestonk_terminal/cryptocurrency/technical_analysis/ta_controller.py
@@ -10,15 +10,14 @@ from colorama import Style
 import pandas as pd
 from prompt_toolkit.completion import NestedCompleter
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     EXPORT_BOTH_RAW_DATA_AND_FIGURES,
-    menu_decorator,
     parse_known_args_and_warn,
     check_positive_list,
     check_positive,
     valid_date,
-    try_except,
     system_clear,
 )
 from gamestonk_terminal.menu import session
