+++ gamestonk_terminal/economy/fred/fred_controller.py
@@ -7,15 +7,14 @@ from typing import List, Union, Dict
 from rich.console import Console
 from prompt_toolkit.completion import NestedCompleter
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.economy.fred import fred_view, fred_model
 from gamestonk_terminal.helper_funcs import (
     EXPORT_BOTH_RAW_DATA_AND_FIGURES,
     check_positive,
-    menu_decorator,
     parse_known_args_and_warn,
     valid_date,
-    try_except,
     system_clear,
 )
 from gamestonk_terminal.menu import session
