+++ gamestonk_terminal/economy/fred/prediction/pred_controller.py
@@ -2,22 +2,20 @@
 __docformat__ = "numpy"
 
 import argparse
-import difflib
 from typing import List, Union, Dict
 import logging
 
 from prompt_toolkit.completion import NestedCompleter
 from rich.console import Console
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
-    get_flair,
     parse_known_args_and_warn,
     check_positive,
     valid_date,
     get_next_stock_market_days,
     EXPORT_ONLY_FIGURES_ALLOWED,
-    try_except,
     system_clear,
 )
 from gamestonk_terminal.menu import session
@@ -35,6 +33,7 @@ from gamestonk_terminal.common.prediction_techniques import (
 )
 from gamestonk_terminal.economy.fred import fred_model
 
+# pylint: disable=W0613
 
 logger = logging.getLogger(__name__)
 t_console = Console()
@@ -868,80 +867,6 @@ Models:
             )
 
 
+@menu_decorator("/economy/fred/pred/", PredictionTechniquesController)
 def menu(series: Dict, queue: List[str] = None):
     """Prediction Techniques Menu"""
-
-    pred_controller = PredictionTechniquesController(series, queue)
-    an_input = "HELP_ME"
-
-    while True:
-        # There is a command in the queue
-        if pred_controller.queue and len(pred_controller.queue) > 0:
-            # If the command is quitting the menu we want to return in here
-            if pred_controller.queue[0] in ("q", "..", "quit"):
-                t_console.print("")
-                if len(pred_controller.queue) > 1:
-                    return pred_controller.queue[1:]
-                return []
-
-            # Consume 1 element from the queue
-            an_input = pred_controller.queue[0]
-            pred_controller.queue = pred_controller.queue[1:]
-
-            # Print the current location because this was an instruction and we want user to know what was the action
-            if an_input and an_input.split(" ")[0] in pred_controller.CHOICES_COMMANDS:
-                t_console.print(f"{get_flair()} /economy/fred/pred/ $ {an_input}")
-
-        # Get input command from user
-        else:
-            # Display help menu when entering on this menu from a level above
-            if an_input == "HELP_ME":
-                pred_controller.print_help()
-
-            # Get input from user using auto-completion
-            if session and gtff.USE_PROMPT_TOOLKIT and pred_controller.completer:
-                try:
-                    an_input = session.prompt(
-                        f"{get_flair()} /economy/fred/pred/ $ ",
-                        completer=pred_controller.completer,
-                        search_ignore_case=True,
-                    )
-                except KeyboardInterrupt:
-                    # Exit in case of keyboard interrupt
-                    an_input = "exit"
-            # Get input from user without auto-completion
-            else:
-                an_input = input(f"{get_flair()} /economy/fred/pred/ $ ")
-
-        try:
-            # Process the input command
-            pred_controller.queue = pred_controller.switch(an_input)
-
-        except SystemExit:
-            t_console.print(
-                f"\nThe command '{an_input}' doesn't exist on the /economy/fred/pred menu.\n",
-            )
-            similar_cmd = difflib.get_close_matches(
-                an_input.split(" ")[0] if " " in an_input else an_input,
-                pred_controller.CHOICES,
-                n=1,
-                cutoff=0.7,
-            )
-            if similar_cmd:
-                if " " in an_input:
-                    candidate_input = (
-                        f"{similar_cmd[0]} {' '.join(an_input.split(' ')[1:])}"
-                    )
-                    if candidate_input == an_input:
-                        an_input = ""
-                        pred_controller.queue = []
-                        t_console.print("")
-                        continue
-                    an_input = candidate_input
-                else:
-                    an_input = similar_cmd[0]
-
-                t_console.print(f" Replacing by '{an_input}'.")
-                pred_controller.queue.insert(0, an_input)
-            else:
-                t_console.print("")
