+++ gamestonk_terminal/etf/screener/screener_controller.py
@@ -9,12 +9,11 @@ from typing import List, Union
 
 from prompt_toolkit.completion import NestedCompleter
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     EXPORT_ONLY_RAW_DATA_ALLOWED,
-    menu_decorator,
     parse_known_args_and_warn,
-    try_except,
     system_clear,
     check_positive,
 )
