+++ gamestonk_terminal/etf/technical_analysis/ta_controller.py
@@ -10,14 +10,13 @@ from datetime import datetime
 import pandas as pd
 from prompt_toolkit.completion import NestedCompleter
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     EXPORT_BOTH_RAW_DATA_AND_FIGURES,
-    menu_decorator,
     parse_known_args_and_warn,
     check_positive_list,
     check_positive,
-    try_except,
     valid_date,
     system_clear,
 )
