+++ gamestonk_terminal/forex/forex_controller.py
@@ -9,14 +9,13 @@ import pandas as pd
 from colorama import Style
 from prompt_toolkit.completion import NestedCompleter
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.forex import av_view, av_model
 
 from gamestonk_terminal.helper_funcs import (
-    menu_decorator,
     parse_known_args_and_warn,
     valid_date,
-    try_except,
     system_clear,
 )
 from gamestonk_terminal.menu import session
