+++ gamestonk_terminal/mutual_funds/mutual_fund_controller.py
@@ -10,6 +10,7 @@ import pandas as pd
 from prompt_toolkit.completion import NestedCompleter
 from rich import console
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     EXPORT_ONLY_FIGURES_ALLOWED,
@@ -17,10 +18,8 @@ from gamestonk_terminal.helper_funcs import (
     EXPORT_BOTH_RAW_DATA_AND_FIGURES,
     check_non_negative_float,
     check_positive,
-    menu_decorator,
     parse_known_args_and_warn,
     system_clear,
-    try_except,
     valid_date,
 )
 from gamestonk_terminal.menu import session
