+++ gamestonk_terminal/portfolio/brokers/ally/ally_controller.py
@@ -5,13 +5,12 @@ import argparse
 from typing import List, Union
 from prompt_toolkit.completion import NestedCompleter
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.menu import session
 from gamestonk_terminal.portfolio.brokers.ally import ally_view
 from gamestonk_terminal.helper_funcs import (
-    menu_decorator,
     parse_known_args_and_warn,
-    try_except,
     system_clear,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
 )
