+++ gamestonk_terminal/portfolio/brokers/bro_controller.py
@@ -8,7 +8,8 @@ from prompt_toolkit.completion import NestedCompleter
 
 
 from gamestonk_terminal import feature_flags as gtff
-from gamestonk_terminal.helper_funcs import system_clear, menu_decorator
+from gamestonk_terminal.decorators import menu_decorator
+from gamestonk_terminal.helper_funcs import system_clear
 from gamestonk_terminal.menu import session
 
 from gamestonk_terminal.portfolio.brokers.ally import ally_controller
