+++ gamestonk_terminal/portfolio/brokers/robinhood/robinhood_controller.py
@@ -11,10 +11,9 @@ from gamestonk_terminal.portfolio.brokers.robinhood import (
     robinhood_view,
     robinhood_model,
 )
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal.helper_funcs import (
-    menu_decorator,
     parse_known_args_and_warn,
-    try_except,
     system_clear,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
 )
