+++ gamestonk_terminal/portfolio/portfolio_analysis/pa_controller.py
@@ -6,10 +6,9 @@ import os
 
 import pandas as pd
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal.helper_funcs import (
-    menu_decorator,
     parse_known_args_and_warn,
-    try_except,
     system_clear,
 )
 from gamestonk_terminal.portfolio.portfolio_analysis import (
