+++ gamestonk_terminal/portfolio/portfolio_controller.py
@@ -11,10 +11,9 @@ from datetime import datetime
 from prompt_toolkit.completion import NestedCompleter
 import pandas as pd
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
-    menu_decorator,
-    try_except,
     valid_date,
     check_positive_float,
     system_clear,
