+++ gamestonk_terminal/portfolio/portfolio_optimization/po_controller.py
@@ -6,11 +6,10 @@ from typing import List, Union
 
 from prompt_toolkit.completion import NestedCompleter
 from gamestonk_terminal import feature_flags as gtff
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal.helper_funcs import (
-    menu_decorator,
     parse_known_args_and_warn,
     check_non_negative,
-    try_except,
     system_clear,
 )
 from gamestonk_terminal.menu import session
