+++ gamestonk_terminal/stocks/backtesting/bt_controller.py
@@ -8,13 +8,12 @@ import matplotlib as mpl
 import pandas as pd
 from prompt_toolkit.completion import NestedCompleter
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     check_non_negative_float,
     check_positive,
     parse_known_args_and_warn,
-    try_except,
-    menu_decorator,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
     system_clear,
     valid_date,
