+++ gamestonk_terminal/stocks/behavioural_analysis/ba_controller.py
@@ -7,16 +7,16 @@ from datetime import datetime, timedelta
 import textwrap
 from prompt_toolkit.completion import NestedCompleter
 from colorama import Style
+
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     EXPORT_BOTH_RAW_DATA_AND_FIGURES,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
     parse_known_args_and_warn,
     check_int_range,
-    menu_decorator,
     valid_date,
     check_positive,
-    try_except,
     system_clear,
 )
 from gamestonk_terminal.menu import session
