+++ gamestonk_terminal/stocks/comparison_analysis/ca_controller.py
@@ -9,13 +9,12 @@ import yfinance as yf
 from colorama import Style
 from prompt_toolkit.completion import NestedCompleter
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     check_non_negative,
     check_positive,
-    menu_decorator,
     parse_known_args_and_warn,
-    try_except,
     system_clear,
     valid_date,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
