+++ gamestonk_terminal/stocks/dark_pool_shorts/dps_controller.py
@@ -7,15 +7,14 @@ from datetime import datetime, timedelta
 from colorama import Style
 import pandas as pd
 from prompt_toolkit.completion import NestedCompleter
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.menu import session
 from gamestonk_terminal.helper_funcs import (
     parse_known_args_and_warn,
     check_positive,
-    menu_decorator,
     valid_date,
     check_int_range,
-    try_except,
     system_clear,
     EXPORT_BOTH_RAW_DATA_AND_FIGURES,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
