+++ gamestonk_terminal/stocks/discovery/disc_controller.py
@@ -6,15 +6,14 @@ from datetime import datetime
 from typing import List, Union
 from prompt_toolkit.completion import NestedCompleter
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     parse_known_args_and_warn,
     check_non_negative,
     check_positive,
     check_int_range,
-    try_except,
     system_clear,
-    menu_decorator,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
     valid_date,
 )
