+++ gamestonk_terminal/stocks/fundamental_analysis/fa_controller.py
@@ -8,6 +8,7 @@ from typing import List, Union
 from prompt_toolkit.completion import NestedCompleter
 from colorama import Style
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal.stocks.fundamental_analysis.financial_modeling_prep import (
     fmp_controller,
     fmp_view,
@@ -24,10 +25,8 @@ from gamestonk_terminal.stocks.fundamental_analysis import (
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     EXPORT_ONLY_RAW_DATA_ALLOWED,
-    menu_decorator,
     parse_known_args_and_warn,
     check_positive,
-    try_except,
     system_clear,
     valid_date,
 )
