+++ gamestonk_terminal/stocks/fundamental_analysis/financial_modeling_prep/fmp_controller.py
@@ -5,16 +5,15 @@ import argparse
 from typing import List, Union
 from prompt_toolkit.completion import NestedCompleter
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal.stocks.fundamental_analysis.financial_modeling_prep import (
     fmp_view,
 )
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
-    menu_decorator,
     parse_known_args_and_warn,
     check_positive,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
-    try_except,
     system_clear,
 )
 from gamestonk_terminal.menu import session
