+++ gamestonk_terminal/stocks/government/gov_controller.py
@@ -6,14 +6,14 @@ from datetime import datetime, timedelta
 from typing import List, Union
 from colorama import Style
 from prompt_toolkit.completion import NestedCompleter
+
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.menu import session
 from gamestonk_terminal.stocks.government import quiverquant_view
 from gamestonk_terminal.helper_funcs import (
     parse_known_args_and_warn,
-    menu_decorator,
     check_positive,
-    try_except,
     system_clear,
     valid_date,
 )
