+++ gamestonk_terminal/stocks/insider/insider_controller.py
@@ -8,12 +8,11 @@ from typing import List, Union
 import pandas as pd
 from colorama import Style
 from prompt_toolkit.completion import NestedCompleter
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal.helper_funcs import (
     EXPORT_ONLY_RAW_DATA_ALLOWED,
-    menu_decorator,
     parse_known_args_and_warn,
     check_positive,
-    try_except,
     system_clear,
 )
 from gamestonk_terminal.menu import session
