+++ gamestonk_terminal/stocks/options/payoff_controller.py
@@ -6,11 +6,10 @@ from typing import List, Dict, Union
 from prompt_toolkit.completion import NestedCompleter
 from colorama import Style
 from gamestonk_terminal import feature_flags as gtff
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal.helper_funcs import (
     check_non_negative,
-    menu_decorator,
     parse_known_args_and_warn,
-    try_except,
     system_clear,
 )
 from gamestonk_terminal.menu import session
