+++ gamestonk_terminal/stocks/options/pricing_controller.py
@@ -8,10 +8,9 @@ from tabulate import tabulate
 from prompt_toolkit.completion import NestedCompleter
 
 from gamestonk_terminal import feature_flags as gtff
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal.helper_funcs import (
-    menu_decorator,
     parse_known_args_and_warn,
-    try_except,
     system_clear,
 )
 from gamestonk_terminal.menu import session
