+++ gamestonk_terminal/stocks/options/screener_controller.py
@@ -7,14 +7,13 @@ from typing import List, Union
 
 from colorama import Style
 from prompt_toolkit.completion import NestedCompleter
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     EXPORT_ONLY_RAW_DATA_ALLOWED,
     check_positive,
-    menu_decorator,
     parse_known_args_and_warn,
     system_clear,
-    try_except,
 )
 from gamestonk_terminal.menu import session
 from gamestonk_terminal.portfolio.portfolio_optimization import po_controller
