+++ gamestonk_terminal/stocks/quantitative_analysis/qa_controller.py
@@ -9,6 +9,7 @@ import numpy as np
 import pandas as pd
 from rich.console import Console
 from prompt_toolkit.completion import NestedCompleter
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal.common.quantitative_analysis import (
     qa_view,
     rolling_view,
@@ -18,11 +19,9 @@ from gamestonk_terminal.stocks import stocks_helper
 from gamestonk_terminal.helper_funcs import (
     EXPORT_ONLY_RAW_DATA_ALLOWED,
     EXPORT_ONLY_FIGURES_ALLOWED,
-    menu_decorator,
     check_positive,
     check_proportion_range,
     parse_known_args_and_warn,
-    try_except,
     system_clear,
     valid_date,
 )
