+++ gamestonk_terminal/stocks/screener/screener_controller.py
@@ -9,15 +9,14 @@ from typing import List, Union
 
 from colorama import Style
 from prompt_toolkit.completion import NestedCompleter
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     EXPORT_BOTH_RAW_DATA_AND_FIGURES,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
     check_positive,
-    menu_decorator,
     parse_known_args_and_warn,
     system_clear,
-    try_except,
     valid_date,
 )
 from gamestonk_terminal.menu import session
