+++ gamestonk_terminal/stocks/sector_industry_analysis/sia_controller.py
@@ -7,11 +7,10 @@ from typing import List, Union
 import yfinance as yf
 from colorama import Style
 from prompt_toolkit.completion import NestedCompleter
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal.helper_funcs import (
     EXPORT_BOTH_RAW_DATA_AND_FIGURES,
-    menu_decorator,
     parse_known_args_and_warn,
-    try_except,
     system_clear,
     check_positive,
     check_proportion_range,
