+++ gamestonk_terminal/stocks/technical_analysis/ta_controller.py
@@ -9,16 +9,15 @@ from datetime import datetime, timedelta
 import pandas as pd
 from prompt_toolkit.completion import NestedCompleter
 
+from gamestonk_terminal.decorators import try_except, menu_decorator
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import (
     EXPORT_BOTH_RAW_DATA_AND_FIGURES,
     EXPORT_ONLY_FIGURES_ALLOWED,
     EXPORT_ONLY_RAW_DATA_ALLOWED,
-    menu_decorator,
     parse_known_args_and_warn,
     check_positive_list,
     check_positive,
-    try_except,
     valid_date,
     system_clear,
 )
